changed the id to BigInt (long)

+++ src/main/java/de/dan/hobby/moisql/table/Inserter.java
@@ -1,6 +1,7 @@
 package de.dan.hobby.moisql.table;
 
 import de.dan.hobby.moisql.datatype.IDataType;
+import de.dan.hobby.moisql.datatype.numeric.BigInt;
 import de.dan.hobby.moisql.datatype.numeric.Int;
 import de.dan.hobby.moisql.tree.BPTree;
 
@@ -30,7 +31,7 @@ public class Inserter {
         throw new IllegalArgumentException("datatypes are not consistent with table");
       }
 
-      Int tempId = (Int)tableRow[0];
+      BigInt tempId = (BigInt)tableRow[0];
       if(tree.findRow(tempId.getValue()) == null){
         tree.insertRow(tableRow);
       }
