+++ src/main/java/de/dan/hobby/moisql/tree/Node.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public abstract class Node {
 
-  List<Integer> keys = new ArrayList<>();
+  List<Long> keys = new ArrayList<>();
 
   Node parent;
